cancelled replied to string and fixed session condition

+++ view/forum/detailsTopic.php
@@ -38,9 +38,6 @@ if (isset($posts)) {
     foreach ($posts as $post) { ?>
         <div class="d-flex bg-light py-2 my-5 border border-dark-subtle rounded"
             style="min-width: 35%; max-width: fit-content;">
-            <?php if (!is_null($post->getPost())) { ?>
-                <p>En réponse à <?= $post->getPost()->getUser() ?></p>
-            <?php } ?>
 
             <div class="m-2"><img class="img-thumbnail rounded w-5" style="object-fit-cover"
                     src="public/img/avatar/<?= $post->getUser()->getAvatar() ?>" alt="">
@@ -77,19 +74,21 @@ if (isset($posts)) {
                             </div>
                         </form>
                     </div>
+                <?php }
+                if (isset($_SESSION['user'])) { ?>
+                    <div class="d-flex">
+                        <!-- if the current user is an admin or the author of this topic -->
+                        <?php if ($_SESSION['user'] == $post->getUser() || App\Session::isAdmin()) { ?>
+                            <a class="delete-btn btn btn-outline-danger m-2"
+                                href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer</i></a>
+                        <?php }
+                        // if the current user is the author of this post
+                        if ($_SESSION['user'] == $post->getUser()) { ?>
+                            <a class="btn btn-outline-dark m-2"
+                                href="index.php?ctrl=forum&action=updatePost&id=<?= $post->getId() ?>">Modifier</a>
+                        <?php } ?>
+                    </div>
                 <?php } ?>
-                <div class="d-flex">
-                    <!-- if the current user is an admin or the author of this topic -->
-                    <?php if ($_SESSION['user'] == $post->getUser() || App\Session::isAdmin()) { ?>
-                        <a class="delete-btn btn btn-outline-danger m-2"
-                            href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer</i></a>
-                    <?php }
-                    // if the current user is the author of this post
-                    if ($_SESSION['user'] == $post->getUser()) { ?>
-                        <a class="btn btn-outline-dark m-2"
-                            href="index.php?ctrl=forum&action=updatePost&id=<?= $post->getId() ?>">Modifier</a>
-                    <?php } ?>
-                </div>
             </div>
         </div>
     <?php }
